Add argument parsing to create kube pod script.

+++ airbyte-tests/bin/create-kube-pod.sh
@@ -1,12 +1,25 @@
 #!/usr/bin/env bash
 
-# This script requires gdate and num.
-# To install gdata: brew install coreutils
-# To install num: https://github.com/numcommand/num
+create_only=0
+
+while getopts 'c:h' opt; do
+  case "$opt" in
+    a)
+      echo "Running in create-only mode.."
+      create_only=1
+      ;;
+
+    ?|h)
+      echo "Usage: $(basename $0) [-c]"
+      exit 1
+      ;;
+  esac
+done
+shift "$(($OPTIND -1))"
 
 # watch api and pipe to file
 curl -sN http://127.0.0.1:8001/api/v1/namespaces/default/pods\?watch\=1 | \
-  jq -c 'select(.object.status.phase == "Running") | { "name": .object.metadata.name, "namespace": .object.metadata.namespace, "creationTimestamp": .object.metadata.creationTimestamp, "phase": .object.status.phase, "conditions": [ .object.status.conditions | .[] | select(.status == "True") | {(.type): .lastTransitionTime} ] | add } | select(.conditions.Ready)' >> ./data.yaml &
+  jq -c 'select(.object.status.phase == "Running") | { "name": .object.metadata.name, "namespace": .object.metadata.namespace, "creationTimestamp": .object.metadata.creationTimestamp, "phase": .object.status.phase, "conditions": [ .object.status.conditions | .[] | select(.status == "True") | {(.type): .lastTransitionTime} ] | add } | select(.conditions.Ready)' >> ./data-gke-1.21.txt &
 
 
 i=1
@@ -15,16 +28,21 @@ do
   echo "creating pod number: $i"
   kubectl create -f ./test-pod.yaml
 
-  # wait till pod is in complete
-  while [ "$(kubectl get pod | grep -c 'Compl')" -eq 0 ]
-  do
-    echo 'wait for pod to finish..'
-    sleep 2
-  done
+  if [ $create_only -eq 0 ]
+  then
+      echo "running in clean up mode.."
+      
+      # wait till pod is in complete
+      while [ "$(kubectl get pod | grep -c 'Compl')" -eq 0 ]
+      do
+        echo 'wait for pod to finish..'
+        sleep 2
+      done
 
-  # remove old pod to not mess with results; sleep to let Kube cluster rest
-  kubectl get pod | grep Comp | awk '{print $1}' | xargs kubectl delete pod
-  sleep 1;
+      # remove old pod to not mess with results; sleep to let Kube cluster rest
+      kubectl get pod | grep Comp | awk '{print $1}' | xargs kubectl delete pod
+      sleep 1;
 
+  fi
   i=$((i+1))
 done
